Refactored and moved cleaning transformations
